Created route group for (auth)... so cleaner
