+++ src/app/(authenticated)/layout.tsx
@@ -1,0 +1,16 @@
+import Navigation from '@auth/navigation'
+import { Breadcrumbs } from '@elements/breadcrumbs'
+
+export default async function RootLayout({
+    children,
+}: {
+    children: React.ReactNode
+}) {
+    return (
+        // @ts-expect-error
+        <Navigation>
+            <Breadcrumbs />
+            {children}
+        </Navigation>
+    )
+}
