+++ src/app/(authenticated)/protocols/[id]/[section]/page.tsx
