+++ src/app/(authenticated)/protocols/[id]/layout.tsx
