+++ src/app/(authenticated)/protocols/[id]/loading.tsx
