+++ src/app/(authenticated)/protocols/[id]/page.tsx
