+++ src/app/(authenticated)/protocols/loading.tsx
