+++ src/app/page.tsx
@@ -1,5 +1,9 @@
+import { SignIn } from '@auth/sign-in'
+import { getServerSession } from 'next-auth'
 import { redirect } from 'next/navigation'
 
 export default async function Page() {
-    redirect('/protocols')
+    const session = await getServerSession()
+    if (session) redirect('/protocols')
+    return <SignIn />
 }
